# docs/swagger.yaml
basePath: /
definitions:
...
  ent.Album:
    properties:
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      description:
        description: 相册分类描述
        type: string
      id:
        description: ID of the ent.
        type: integer
      name:
        description: 相册分类名称
        type: string
      sort:
        description: 相册分类排序
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
  ent.AlbumPhoto:
    properties:
      album_id:
        description: 相册ID
        type: integer
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      description:
        description: 相片描述
        type: string
      id:
        description: ID of the ent.
        type: integer
      image_url:
        description: 图片地址
        type: string
      name:
        description: 相片名称
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      view_count:
        description: 查看次数
        type: integer
    type: object
...
  model.AlbumCreateReq:
    properties:
      description:
        type: string
      name:
        type: string
      sort:
        type: integer
    required:
    - name
    type: object
  model.AlbumPhotoCreateReq:
    properties:
      album_id:
        type: integer
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
    required:
    - album_id
    - image_url
    type: object
  model.AlbumPhotoUpdateReq:
    properties:
      album_id:
        type: integer
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
    type: object
  model.AlbumUpdateReq:
    properties:
      description:
        type: string
      name:
        type: string
      sort:
        type: integer
    type: object
...
  model.HttpError:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  model.HttpSuccess:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  model.PageResult-ent_Album:
    properties:
      records:
        items:
          $ref: '#/definitions/ent.Album'
        type: array
      total:
        type: integer
    type: object
  model.PageResult-ent_AlbumPhoto:
    properties:
      records:
        items:
          $ref: '#/definitions/ent.AlbumPhoto'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:13000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /api/v1/album-photo/create:
    post:
      consumes:
      - application/json
      description: 创建一个新的相册照片
      parameters:
      - description: 相册照片信息
        in: body
        name: album_photo
        required: true
        schema:
          $ref: '#/definitions/model.AlbumPhotoCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.AlbumPhotoCreateReq'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 创建相册照片
      tags:
      - albumPhotos
  /api/v1/album-photo/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定相册照片
      parameters:
      - description: 相册照片ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 删除相册照片
      tags:
      - albumPhotos
  /api/v1/album-photo/list:
    get:
      consumes:
      - application/json
      description: 查询所有相册照片
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ent.AlbumPhoto'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询相册照片列表
      tags:
      - albumPhotos
  /api/v1/album-photo/page:
    get:
      consumes:
      - application/json
      description: 查询所有相册照片分页列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PageResult-ent_AlbumPhoto'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询相册照片分页列表
      tags:
      - albumPhotos
  /api/v1/album-photo/query/{id}:
    get:
      consumes:
      - application/json
      description: 查询指定相册照片的信息
      parameters:
      - description: 相册照片ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.AlbumPhoto'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询相册照片
      tags:
      - albumPhotos
  /api/v1/album-photo/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定相册照片的信息
      parameters:
      - description: 相册照片ID
        in: path
        name: id
        required: true
        type: string
      - description: 相册照片信息
        in: body
        name: album_photo
        required: true
        schema:
          $ref: '#/definitions/model.AlbumPhotoUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.AlbumPhotoUpdateReq'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 更新相册照片
      tags:
      - albumPhotos
  /api/v1/album/create:
    post:
      consumes:
      - application/json
      description: 创建一个新相册
      parameters:
      - description: 相册创建请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.AlbumCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.AlbumCreateReq'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 创建相册
      tags:
      - album
  /api/v1/album/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定相册
      parameters:
      - description: 相册ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 删除相册
      tags:
      - album
  /api/v1/album/list:
    get:
      consumes:
      - application/json
      description: 查询所有相册
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ent.Album'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询相册列表
      tags:
      - album
  /api/v1/album/page:
    get:
      consumes:
      - application/json
      description: 查询相册列表分页
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PageResult-ent_Album'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询相册列表分页
      tags:
      - album
  /api/v1/album/query/{id}:
    get:
      consumes:
      - application/json
      description: 查询指定相册的信息
      parameters:
      - description: 相册ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.Album'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询相册
      tags:
      - album
  /api/v1/album/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定相册的信息
      parameters:
      - description: 相册ID
        in: path
        name: id
        required: true
        type: string
      - description: 相册更新请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.AlbumUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.AlbumUpdateReq'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 更新相册
      tags:
      - album
...
swagger: "2.0"